Implement restore using POST /v1/import-hierarchy
Without this patch there is no ability to restore a backup taken with the
`backup subcommand`.  This patch implements restore functionality using the
[import-hierarchy](https://docs.puppet.com/pe/2016.1/nc_import-hierarchy.html)
API endpoint.  Input data, either from STDIN or from a file is streamed to the
service using HTTP 1.1 chunked encoding.  This should make restoration of very
large node classification backup files robust and effective.

Note, backups taken with the backup subcommand should only be restored to the
same PE infrastructure they were taken on.  This is because the backup data
itself contains embedded fully qualified domain names.  If a backup from one
infrastructure is restored to another, e.g. a Monolithic Master designated as a
secondary, then the PE infrastructure itself will fail to operate due to
mis-configured service URL's and hostnames.

A future enhancement will implement a transformation subcommand to facilitate
restoring a backup to an arbitrary PE Monolithic master.
